Reorganize settingsSchema import to avoid circular dependency
