+++ apps/web-client/src/lib/forms/SettingsForm.svelte
@@ -4,7 +4,7 @@
 	import compare from "just-compare";
 
 	import { Input } from "$lib/components";
-	import type { SettingsSchema } from "$lib/forms/schemas";
+	import type { SettingsSchema } from "$lib/schemas";
 
 	export let data: SuperValidated<SettingsSchema>;
 	export let options: FormOptions<SettingsSchema>;
