+++ apps/web-client/src/lib/schemas/index.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+import type { Infer } from "sveltekit-superforms";
+
+export type SettingsSchema = Infer<typeof settingsSchema>;
+export const settingsSchema = z.object({
+	couchUrl: z.string(),
+	labelPrinterUrl: z.string(),
+	receiptPrinterUrl: z.string()
+});
