+++ apps/web-client/src/routes/settings/+page.ts
@@ -1,16 +1,15 @@
 import { superValidate } from "sveltekit-superforms/client";
 import { zod } from "sveltekit-superforms/adapters";
 
-import { settingsSchema } from "$lib/forms/schemas";
-import { settingsStore } from "$lib/stores";
-
 import type { PageLoad } from "./$types";
 import { get } from "svelte/store";
+import { settingsStore } from "$lib/stores/app";
+import { settingsSchema } from "$lib/schemas";
 
 export const load: PageLoad = async () => {
 	const settingsData = get(settingsStore);
 
-	const form = await superValidate(settingsData, zod(settingsSchema));
+	const form = await superValidate(settingsData.defaultSettings, zod(settingsSchema));
 
 	return {
 		form
